Moved all new components into components folder
